[python] Some improvements
