+++ sc-kpm/python/_definitions/sc/__init__.py
@@ -1,0 +1,2 @@
+from .sc import ScMemoryContext
+from .sc_class import *
