+++ sc-kpm/python/_definitions/sc/sc_class/__init__.py
@@ -1,0 +1,5 @@
+from .sc_addr import ScAddr
+from .sc_type import ScType
+from .sc_template import ScTemplate, ScTemplateGenParams, ScTemplateGenResult, ScTemplateSearchResult, ScTemplateSearchResultItem
+from .sc_link_content import ScLinkContent
+from .sc_iterator import ScIterator3, ScIterator5
