+++ sc-kpm/python/_definitions/sc/sc_class/sc_iterator.py
@@ -1,0 +1,22 @@
+from sc_addr import ScAddr
+
+class ScIterator3:
+  def Next(self) -> bool:
+    return False
+
+  def IsValid(self) -> bool:
+    return False
+
+  def Get(self, idx: int) -> ScAddr:
+    return ScAddr()
+
+
+class ScIterator5:
+  def Next(self) -> bool:
+    return False
+
+  def IsValid(self) -> bool:
+    return False
+
+  def Get(self, idx: int) -> ScAddr:
+    return ScAddr()
